+++ core/modules/modScrumProject.class.php
@@ -64,7 +64,9 @@ class modScrumProject extends DolibarrModules
 		$this->editor_name = 'ATM Consulting';
 		$this->editor_url = 'www.atm-consulting.fr';
 		// Possible values for version are: 'development', 'experimental', 'dolibarr', 'dolibarr_deprecated' or a version string like 'x.y.z'
-		$this->version = '1.2.3';
+
+		$this->version = '1.3.0';
+
 		// Url to the file with your last numberversion of this module
 		//$this->url_last_version = 'http://www.example.com/versionmodule.txt';