+++ lib/scrumproject_scrumsprint.lib.php
@@ -86,6 +86,16 @@ function scrumsprintPrepareHead($object)
 	$h++;
 
 
+	$kanban = $object->getKanbanId();
+	if ($kanban > 0) {
+		$head[$h][0] = dol_buildpath('/scrumproject/scrumkanban_view.php', 1) . '?id=' . $kanban;
+		$head[$h][1] = $langs->trans('Kanban');
+		$head[$h][2] = 'kanban';
+		$h++;
+	} else {
+		setEventMessage($langs->trans('ScrumKanbanNotDefine',$object->label), 'warnings');
+	}
+
 	// Show more tabs from modules
 	// Entries must be declared in modules descriptor with line
 	//$this->tabs = array(